edits

+++ CSC 1300 Programs/ijhanna42_program4/ijhanna42_prog4.h
@@ -11,7 +11,7 @@ const int maxCapacity = 75;
 
 //Cost structure
 struct Cost {
-    double numCareHours;
+    int numCareHours;
     double costPerHour;
     double foodCost;
     double materialCost;
